[python] extract converter_expr.cpp from python_converter.cpp
Move the get_expr dispatcher and its 7 sibling expression methods of
python_converter into src/python-frontend/converter/converter_expr.cpp:
get_expr, make_char_array_expr, get_literal, is_bytes_literal,
get_lambda_expr, get_tuple_expr, build_slice_object,
build_slice_from_args, plus get_expression_type and make_slice_struct_expr.

python_converter.cpp now stands at ~660 lines: constructor, destructor,
convert(), create_builtin_symbols, import handling, and small accessors.
Total file count under src/python-frontend/converter/ is 11.

Pure code movement — no logic change.
