+++ src/python-frontend/CMakeLists.txt
@@ -17,6 +17,7 @@ add_library(pythonfrontend STATIC
             converter/converter_class.cpp
             converter/converter_compare.cpp
             converter/converter_dunder.cpp
+            converter/converter_expr.cpp
             converter/converter_funcall.cpp
             converter/converter_funcdef.cpp
             converter/converter_stmt.cpp
